Implementado fluxo do JWT Token e suas respectivas validações, sendo que um usuário poderá ter mais de um perfil

+++ src/main/java/br/com/nlw/events/application/usecases/invite/gateway/GetInviteHintsByPrettyNameAndUserIdUseCase.java
@@ -1,5 +1,5 @@
 package br.com.nlw.events.application.usecases.invite.gateway;
 
 public interface GetInviteHintsByPrettyNameAndUserIdUseCase {
-    Integer execute(final String eventPrettyName, final Integer userId);
+    Long execute(final String eventPrettyName, final Long userId);
 }
