+++ src/main/java/br/com/nlw/events/application/usecases/invite/gateway/IncrementInviteHintByPrettyNameAndUserIdUseCase.java
@@ -1,5 +1,5 @@
 package br.com.nlw.events.application.usecases.invite.gateway;
 
 public interface IncrementInviteHintByPrettyNameAndUserIdUseCase {
-    void execute(final String eventPrettyName, final Integer userId);
+    void execute(final String eventPrettyName, final Long userId);
 }
