+++ src/main/java/br/com/nlw/events/application/usecases/user/gateway/FindUserByIdUseCase.java
@@ -1,9 +1,7 @@
 package br.com.nlw.events.application.usecases.user.gateway;
 
-import br.com.nlw.events.domain.model.User;
-
-import java.util.UUID;
+import br.com.nlw.events.domain.models.User;
 
 public interface FindUserByIdUseCase {
-    User execute(final Integer userId);
+    User execute(final Long userId);
 }
