+++ src/main/java/br/com/nlw/events/application/usecases/user/impl/FindUserByIdUseCaseImpl.java
@@ -1,11 +1,9 @@
 package br.com.nlw.events.application.usecases.user.impl;
 
 import br.com.nlw.events.application.usecases.user.gateway.FindUserByIdUseCase;
-import br.com.nlw.events.domain.model.User;
+import br.com.nlw.events.domain.models.User;
 import br.com.nlw.events.interfaces.gateway.database.UserGateway;
 import org.springframework.stereotype.Service;
-
-import java.util.UUID;
 
 @Service
 public class FindUserByIdUseCaseImpl implements FindUserByIdUseCase {
@@ -17,7 +15,7 @@ public class FindUserByIdUseCaseImpl implements FindUserByIdUseCase {
     }
 
     @Override
-    public User execute(final Integer userId) {
+    public User execute(final Long userId) {
         if (userId == null) {
             return null;
         }
