+++ src/main/java/br/com/nlw/events/interfaces/dtos/subscription/SubscriptionResponseDTO.java
@@ -1,4 +1,4 @@
 package br.com.nlw.events.interfaces.dtos.subscription;
 
-public record SubscriptionResponseDTO(Integer subscriptionNumber, String designation) {
+public record SubscriptionResponseDTO(Long subscriptionNumber, String designation) {
 }
